% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRBQFR1.R
\name{frbqfr1_model}
\alias{frbqfr1_model}
\title{French base quarter front month trade signal}
\usage{
frbqfr1_model(frq, gas)
}
\arguments{
\item{frq}{an xts object. this should be the day closing prices of the front
quarter contract of the french base quarter
contract.https://www.eex.com/en/market-data/power/futures/french-futures#!/2017/09/26}

\item{gas}{an xts object. this should be the day closing prices of the front
quarter contract of the gasoil  month contract.LOW SULPHUR GASOIL FUTURES
https://www.theice.com/products/34361119/Low-Sulphur-Gasoil-Futures/specs}
}
\value{
1 or -1 if the signal is buy or sell resp.
}
\description{
French base quarter front month trade signal
}
\examples{
frbqfr1_model (frq, gas)
}
