% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model23}
\alias{model23}
\title{The data files containig the model}
\format{an rdata object}
\source{
personal
}
\usage{
data(model23)
}
\description{
The data files containig the model
}
\examples{
data(model23)

}
\keyword{models}
